Biological-Image-Processing-and-Informatics/homework03/requirements.txt
albumentations~=1.3.1
opencv-python~=4.9.0.80
matplotlib~=3.5.3
numpy~=1.21.5
PyYAML~=6.0.1
scikit-learn~=1.0.2
tqdm~=4.66.1
pandas~=1.3.5